requirements.txt
fastapi==0.135.3
uvicorn==0.42.0
sqlalchemy
alembic
psycopg2-binary==2.9.11
pydantic
python-jose==3.5.0
bcrypt==4.0.1
passlib==1.7.4
anthropic==0.86.0
python-dotenv